Reflect code review
- Remove variance for function on return type of ServiceConfig#requestIdGenerator()
- Use BlockingWebClient in the test
- Make separate test for requestId
