+++ core/src/main/java/com/linecorp/armeria/server/HttpServerHandler.java
@@ -679,4 +679,3 @@ final class HttpServerHandler extends ChannelInboundHandlerAdapter implements Ht
         }
     }
 }
-
