+++ core/src/main/java/com/linecorp/armeria/server/VirtualHostBuilder.java
@@ -1051,7 +1051,7 @@ public final class VirtualHostBuilder implements TlsSetters {
 
     /**
      * Sets the {@link Function} which generates a {@link RequestId}.
-     * If not set, the value set via {@link ServerBuilder#requestIdGenerator(Supplier)} is used.
+     * If not set, the value set via {@link ServerBuilder#requestIdGenerator(Function)} is used.
      *
      * @param requestIdGenerator the {@link Function} which generates a {@link RequestId}
      * @see RequestContext#id()
